Changes resulting from using the library in Wire Android
After removing the signals functionality from Wire Android codebase (https://github.com/wireapp/wire-android) it became obvious
that certain changes are required, and other would be nice to try out. There was a main issue with the execution context:
In wire-signals I want ExecutionContext.global to be the default one. In Wire Android there is a custom main execution threadpool
which was used as the default by signals as well. I made Wire Android create that custom threadpool and switch it for default
in wire-android.
Another thing was that we don't know yet what classes should be open for extension. I don't want to make them all open as it may
affect performance. For now I made some ad-hoc decisions. We should clarify it before we release wire-signals 1.0.