Introduce similar naming for wrapping signals and streams over data
Until now, the naming of methods which created signals and event streams from initial data was quite inconsistent.
We had `wrap` to wrap a signal over a stream, but also `future` to wrap a signal over a future, and there was `zip`
to zip one signal with another, but also an `apply` which was doing the same thing, but for zipping event streams
we used `union` instead, and anyway `apply` in signals was used to create cosnt signals from any data.

This PR tries to make sense out of it. `Signal.apply` is now used solely to create `ConstSignal`. It takes only
one parameter. For transformation of a future into a signal or an event stream into a signal, or a signal into
an event stream (even more transformers will be coming in next PRs) we now have `Signal.from` and `EventStream.from`.
And for zipping a few event streams or signals together, we have `Signal.zip` and `EventStream.zip`.

I also used this PR to hide a few subclasses of `Signal` and `EventStream` from API. The user is still able to
create them with methods of respective companion objects, but will not be able to do it directly. This is
an approach I'd like to take with the whole library: smart constructors and special methods in the companion object
instead of direct constructors.