fix: ThrottledSignal sometimes ignores newer values
With the old implementation of ThrottledSignal, if the signal is initialized with a sequence of quick changes,
it is possible that it will store the first value and ignore the others. See https://github.com/wireapp/wire-signals/issues/47.
The new implementation makes sure the last value from the sequence is emited.